requirements.txt
adjango>=4.2.0,<5.0.0
djangorestframework
django-cors-headers
